Paper over odd amqplib behaviour around consumerTag
Rabbit will automatically create a consumerTag for you if you don't pass one in, but the only time amqplib tells you what that tag is is when you get a message back from the queue. It never resolves the `subscribe` promise with it.

Work around that by setting the consumerTag to a uuid which we can be assured is unique, and we can hand back to the caller right away.